// frontend/src/app/services/models/BorrowedBookResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type BorrowedBookResponse = {
    id?: number;
    title?: string;
    authorName?: string;
    isbn?: string;
    rate?: number;
    returned?: boolean;
    returnApproved?: boolean;
    read?: boolean;
    readCount?: number;
    cover?: string;
};